End of session 5

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -59,11 +60,12 @@ namespace AuctionService.Controllers
         }
 
         [HttpPost]
+        [Authorize]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
         {
             var auction = _mapper.Map<Auction>(auctionDto);
 
-            auction.Seller = "test";
+            auction.Seller = User.Identity?.Name ?? "Anonymous";
             _context.Auctions.Add(auction);
 
             var newAuction = _mapper.Map<AuctionDto>(auction);
@@ -81,6 +83,7 @@ namespace AuctionService.Controllers
         }
 
         [HttpPut("{id}")]
+        [Authorize]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto auctionDto)
         {
             var auction = await _context.Auctions.Include(x => x.Item)
@@ -89,6 +92,11 @@ namespace AuctionService.Controllers
             if (auction == null)
             {
                 return NotFound();
+            }
+
+            if (auction.Seller != User.Identity?.Name)
+            {
+                return Forbid();
             }
 
             auction.Item = new Item();
@@ -111,6 +119,7 @@ namespace AuctionService.Controllers
         }
 
         [HttpDelete("{id}")]
+        [Authorize]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
             var auction = await _context.Auctions.FindAsync(id);
@@ -118,6 +127,11 @@ namespace AuctionService.Controllers
             if (auction == null)
             {
                 return NotFound();
+            }
+
+            if (auction.Seller != User.Identity?.Name)
+            {
+                return Forbid();
             }
 
             _context.Auctions.Remove(auction);
